+++ node/src/components/consensus/protocols/zug/wal.rs
@@ -12,7 +12,7 @@ use thiserror::Error;
 use tracing::warn;
 
 use crate::components::consensus::{
-    protocols::simple_consensus::message::{Content, Proposal, SignedMessage},
+    protocols::zug::message::{Content, Proposal, SignedMessage},
     traits::Context,
 };
 
@@ -265,7 +265,7 @@ mod tests {
         // ReadWal when we fail to read it.
         loop {
             // If there are no more entries, we're done
-            if entries.len() == 0 {
+            if entries.is_empty() {
                 break;
             }